Fix 14 bugs from second full codebase review
CRITICAL fixes (crash bugs):
- nfc_fuzzer: Add furi_thread_join before furi_thread_free in worker_free
  (crash on app exit after completed fuzz session)
- nfc_fuzzer: Add furi_thread_join before furi_thread_start in worker_start
  (crash when user runs second fuzz session — thread not in joinable state)
- uart_sniff: Guard refresh event handler against NULL worker (queued
  timer event could fire after worker freed on Back press)

HIGH fixes:
- rayhunter_client: Fix first-poll deadlock — always send poll command
  even before ESP32 has responded, breaking chicken-and-egg where ESP32
  waits for command and app waits for response
- uart_sniff: Fix Back from menu while sniffing — stop worker AND exit
  in one press instead of requiring two (which also triggered NULL deref)

MEDIUM fixes:
- flipperpwn: STRING_DELAY now performs $VARIABLE substitution like
  STRING, STRINGLN, and STRINGLN_DELAY (was typing raw $VAR literally)
- flipperpwn: Remove unnecessary (uint16_t) cast on storage_file_write
  length parameter (fragile truncation risk)
- hid_exfil: Add furi_thread_join before furi_thread_start in worker_start
  (prevents crash if user runs second exfil session)
- spi_flash_dump: Add defensive furi_thread_join before furi_thread_start
  in both spi_worker_start_read and spi_worker_start_verify
- nfc_fuzzer: Add bounds checks on all 4 settings VariableItem callbacks
  to prevent out-of-bounds array access
- rayhunter_client: Remove NULL callback from widget_add_button_element
  in About view (potential crash on button press)
- rogue_ap_detector: Change int8_t RSSI delta to int to prevent overflow
  on extreme RSSI values (both rogue_ap_worker.c and rogue_ap.c)
- subghz_spectrum: Add furi_assert(!worker->running) guard in
  spectrum_worker_free for defensive correctness

LOW fixes:
- badusb_pro: Complete uint16_t -> uint32_t migration for loop variables
  inside flow-control search helpers (find_else_or_endif, find_endif,
  find_end_while, find_matching_while)

Rebuild all 13 FAPs to dist/.

+++ badusb_pro/script_engine.c
@@ -358,7 +358,7 @@ static void press_single_key(uint16_t keycode) {
 /** Skip to matching END_IF or ELSE, handling nesting */
 static uint32_t find_else_or_endif(ScriptEngine* engine, uint32_t from) {
     int depth = 1;
-    for(uint16_t i = from; i < engine->token_count; i++) {
+    for(uint32_t i = from; i < engine->token_count; i++) {
         if(engine->tokens[i].type == TokenIf)
             depth++;
         else if(engine->tokens[i].type == TokenEndIf) {
@@ -374,7 +374,7 @@ static uint32_t find_else_or_endif(ScriptEngine* engine, uint32_t from) {
 /** Skip to matching END_IF from an ELSE block */
 static uint32_t find_endif(ScriptEngine* engine, uint32_t from) {
     int depth = 1;
-    for(uint16_t i = from; i < engine->token_count; i++) {
+    for(uint32_t i = from; i < engine->token_count; i++) {
         if(engine->tokens[i].type == TokenIf)
             depth++;
         else if(engine->tokens[i].type == TokenEndIf) {
@@ -388,7 +388,7 @@ static uint32_t find_endif(ScriptEngine* engine, uint32_t from) {
 /** Skip to matching END_WHILE */
 static uint32_t find_end_while(ScriptEngine* engine, uint32_t from) {
     int depth = 1;
-    for(uint16_t i = from; i < engine->token_count; i++) {
+    for(uint32_t i = from; i < engine->token_count; i++) {
         if(engine->tokens[i].type == TokenWhile)
             depth++;
         else if(engine->tokens[i].type == TokenEndWhile) {
@@ -402,12 +402,12 @@ static uint32_t find_end_while(ScriptEngine* engine, uint32_t from) {
 /** Find the WHILE that matches an END_WHILE (search backwards) */
 static uint32_t find_matching_while(ScriptEngine* engine, uint32_t end_while_idx) {
     int depth = 1;
-    for(int16_t i = (int16_t)(end_while_idx - 1); i >= 0; i--) {
+    for(int32_t i = (int32_t)(end_while_idx - 1); i >= 0; i--) {
         if(engine->tokens[i].type == TokenEndWhile)
             depth++;
         else if(engine->tokens[i].type == TokenWhile) {
             depth--;
-            if(depth == 0) return (uint16_t)i;
+            if(depth == 0) return (uint32_t)i;
         }
     }
     return 0;
